+++ CHANGES.md
@@ -7,9 +7,10 @@
 * Pyrite now uses the N-panel rather than the T-panel, as required for Blender
   2.80 and above.
 * Cleaned up the code (more PEP compliant).
+* Start Over button also deletes Pyrite spheres.
+* Minor improvements to user messages.
 
 1.0.1
 =====
 
-1.0.0
-=====
+* Original version.